Organize `presentation` directory
