+++ pkg/presentation/rest/route.go
@@ -1,7 +1,6 @@
 package rest
 
 import (
-	"drello-api/pkg/presentation/controller"
 	"fmt"
 	"log"
 	"net/http"
@@ -10,6 +9,6 @@ import (
 func HandleRequests() {
 	fmt.Println("Listening on http://127.0.0.1:8080")
 
-	http.HandleFunc("/workspaces", controller.Workspaces)
+	http.HandleFunc("/workspaces", Workspaces)
 	log.Fatal(http.ListenAndServe(":8080", nil))
 }
